requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
thecampy==2.0.0
toml==0.10.2
urllib3==1.26.4
yfinance==0.1.59
